refactor: move util function (now no longer exclusively used in tests)
