+++ flexmeasures/data/tests/test_scheduling_repeated_jobs_fresh_db.py
@@ -4,7 +4,8 @@ from datetime import datetime, timedelta
 
 import pytz
 
-from flexmeasures.data.tests.utils import work_on_rq, exception_reporter
+from flexmeasures.data.tests.utils import exception_reporter
+from flexmeasures.utils.job_utils import work_on_rq
 from flexmeasures.data.services.scheduling import create_scheduling_job
 from flexmeasures.data.models.planning import Scheduler
 from flexmeasures.data.services.scheduling import load_custom_scheduler
